test: separate E2E and unit tests and add AI tests
