+++ backend/tests/e2e/test_interactions.py
@@ -1,1 +1,9 @@
-"""End-to-end tests for the GET /interactions endpoint."""
+import httpx
+
+def test_get_interactions_returns_200(client: httpx.Client) -> None:
+    response = client.get("/interactions/")
+    assert response.status_code == 200
+
+def test_get_interactions_response_is_a_list(client: httpx.Client) -> None:
+    response = client.get("/interactions/")
+    assert isinstance(response.json(), list)
